+++ Src/Infrastructure/Ws.Domain.Models/Entities/Print/Label.cs
@@ -14,7 +14,7 @@ public class Label : EntityBase
     public virtual string BarcodeTop { get; set; } = string.Empty;
     public virtual string BarcodeRight { get; set; } = string.Empty;
     public virtual string BarcodeBottom { get; set; } = string.Empty;
-    public virtual ushort BundleCount { get; set; }
+    public virtual int BundleCount { get; set; }
     public virtual decimal WeightNet { get; set; }
     public virtual decimal WeightTare { get; set; }
     public virtual short Kneading { get; set; }
